{{cookiecutter.project_slug}}/requirements.txt
# Survey Pipeline Requirements
# Core pipeline dependencies

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyodk>=1.0.0
openpyxl>=3.1.0

# Workflow orchestration
prefect>=2.14.0

# Data validation
great-expectations>=0.18.0

# Dashboard
streamlit>=1.28.0
plotly>=5.17.0
folium>=0.15.0
streamlit-folium>=0.15.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0

# Database support (optional)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Email notifications
smtplib-ssl>=1.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Deployment
gunicorn>=21.2.0
supervisor>=4.2.0
